# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Corporate Gifts Inventory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductStockStatus =
  (typeof ProductStockStatus)[keyof typeof ProductStockStatus];

export const ProductStockStatus = {
  in_stock: "in_stock",
  low_stock: "low_stock",
  out_of_stock: "out_of_stock",
} as const;

export interface Product {
  id: number;
  name: string;
  description: string;
  category: string;
  categorySlug: string;
  imageUrl?: string | null;
  basePrice: number;
  discountPrice?: number | null;
  minNegotiationPrice?: number | null;
  isNegotiable: boolean;
  isVisible: boolean;
  stockStatus: ProductStockStatus;
  notes?: string | null;
  bulkPricingNotes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateProductBodyStockStatus =
  (typeof CreateProductBodyStockStatus)[keyof typeof CreateProductBodyStockStatus];

export const CreateProductBodyStockStatus = {
  in_stock: "in_stock",
  low_stock: "low_stock",
  out_of_stock: "out_of_stock",
} as const;

export interface CreateProductBody {
  name: string;
  description: string;
  category: string;
  categorySlug: string;
  imageUrl?: string | null;
  basePrice: number;
  discountPrice?: number | null;
  minNegotiationPrice?: number | null;
  isNegotiable: boolean;
  isVisible: boolean;
  stockStatus: CreateProductBodyStockStatus;
  notes?: string | null;
  bulkPricingNotes?: string | null;
}

export type UpdateProductBodyStockStatus =
  (typeof UpdateProductBodyStockStatus)[keyof typeof UpdateProductBodyStockStatus];

export const UpdateProductBodyStockStatus = {
  in_stock: "in_stock",
  low_stock: "low_stock",
  out_of_stock: "out_of_stock",
} as const;

export interface UpdateProductBody {
  name?: string;
  description?: string;
  category?: string;
  categorySlug?: string;
  imageUrl?: string | null;
  basePrice?: number;
  discountPrice?: number | null;
  minNegotiationPrice?: number | null;
  isNegotiable?: boolean;
  isVisible?: boolean;
  stockStatus?: UpdateProductBodyStockStatus;
  notes?: string | null;
  bulkPricingNotes?: string | null;
}

export interface CategoryWithCount {
  slug: string;
  name: string;
  productCount: number;
}

export interface ProductStats {
  total: number;
  visible: number;
  hidden: number;
  inStock: number;
  lowStock: number;
  outOfStock: number;
  negotiable: number;
  totalCategories: number;
}

export type QuotationStatus =
  (typeof QuotationStatus)[keyof typeof QuotationStatus];

export const QuotationStatus = {
  pending: "pending",
  reviewed: "reviewed",
  responded: "responded",
  closed: "closed",
} as const;

export interface Quotation {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  company?: string | null;
  productIds: number[];
  message: string;
  quantity?: string | null;
  status: QuotationStatus;
  createdAt: string;
}

export interface CreateQuotationBody {
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  company?: string | null;
  productIds: number[];
  message: string;
  quantity?: string | null;
}

export type UpdateQuotationBodyStatus =
  (typeof UpdateQuotationBodyStatus)[keyof typeof UpdateQuotationBodyStatus];

export const UpdateQuotationBodyStatus = {
  pending: "pending",
  reviewed: "reviewed",
  responded: "responded",
  closed: "closed",
} as const;

export interface UpdateQuotationBody {
  status?: UpdateQuotationBodyStatus;
}

export type ListProductsParams = {
  /**
   * Filter by category slug
   */
  category?: string;
  /**
   * Search by name or description
   */
  search?: string;
  /**
   * Filter by visibility
   */
  visible?: boolean;
  /**
   * Filter by negotiable status
   */
  negotiable?: boolean;
};