# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Corporate Gifts Inventory API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all products
 */
export const ListProductsQueryParams = zod.object({
  category: zod.coerce.string().optional().describe("Filter by category slug"),
  search: zod.coerce
    .string()
    .optional()
    .describe("Search by name or description"),
  visible: zod.coerce.boolean().optional().describe("Filter by visibility"),
  negotiable: zod.coerce
    .boolean()
    .optional()
    .describe("Filter by negotiable status"),
});

export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  categorySlug: zod.string(),
  imageUrl: zod.string().nullish(),
  basePrice: zod.number(),
  discountPrice: zod.number().nullish(),
  minNegotiationPrice: zod.number().nullish(),
  isNegotiable: zod.boolean(),
  isVisible: zod.boolean(),
  stockStatus: zod.enum(["in_stock", "low_stock", "out_of_stock"]),
  notes: zod.string().nullish(),
  bulkPricingNotes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Create a new product
 */
export const CreateProductBody = zod.object({
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  categorySlug: zod.string(),
  imageUrl: zod.string().nullish(),
  basePrice: zod.number(),
  discountPrice: zod.number().nullish(),
  minNegotiationPrice: zod.number().nullish(),
  isNegotiable: zod.boolean(),
  isVisible: zod.boolean(),
  stockStatus: zod.enum(["in_stock", "low_stock", "out_of_stock"]),
  notes: zod.string().nullish(),
  bulkPricingNotes: zod.string().nullish(),
});

/**
 * @summary Get a single product
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  categorySlug: zod.string(),
  imageUrl: zod.string().nullish(),
  basePrice: zod.number(),
  discountPrice: zod.number().nullish(),
  minNegotiationPrice: zod.number().nullish(),
  isNegotiable: zod.boolean(),
  isVisible: zod.boolean(),
  stockStatus: zod.enum(["in_stock", "low_stock", "out_of_stock"]),
  notes: zod.string().nullish(),
  bulkPricingNotes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a product
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProductBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().optional(),
  category: zod.string().optional(),
  categorySlug: zod.string().optional(),
  imageUrl: zod.string().nullish(),
  basePrice: zod.number().optional(),
  discountPrice: zod.number().nullish(),
  minNegotiationPrice: zod.number().nullish(),
  isNegotiable: zod.boolean().optional(),
  isVisible: zod.boolean().optional(),
  stockStatus: zod.enum(["in_stock", "low_stock", "out_of_stock"]).optional(),
  notes: zod.string().nullish(),
  bulkPricingNotes: zod.string().nullish(),
});

export const UpdateProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  categorySlug: zod.string(),
  imageUrl: zod.string().nullish(),
  basePrice: zod.number(),
  discountPrice: zod.number().nullish(),
  minNegotiationPrice: zod.number().nullish(),
  isNegotiable: zod.boolean(),
  isVisible: zod.boolean(),
  stockStatus: zod.enum(["in_stock", "low_stock", "out_of_stock"]),
  notes: zod.string().nullish(),
  bulkPricingNotes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a product
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all categories with product counts
 */
export const ListCategoriesResponseItem = zod.object({
  slug: zod.string(),
  name: zod.string(),
  productCount: zod.number(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Get product stats summary
 */
export const GetProductStatsResponse = zod.object({
  total: zod.number(),
  visible: zod.number(),
  hidden: zod.number(),
  inStock: zod.number(),
  lowStock: zod.number(),
  outOfStock: zod.number(),
  negotiable: zod.number(),
  totalCategories: zod.number(),
});

/**
 * @summary List all quotation requests
 */
export const ListQuotationsResponseItem = zod.object({
  id: zod.number(),
  customerName: zod.string(),
  customerEmail: zod.string(),
  customerPhone: zod.string().nullish(),
  company: zod.string().nullish(),
  productIds: zod.array(zod.number()),
  message: zod.string(),
  quantity: zod.string().nullish(),
  status: zod.enum(["pending", "reviewed", "responded", "closed"]),
  createdAt: zod.coerce.date(),
});
export const ListQuotationsResponse = zod.array(ListQuotationsResponseItem);

/**
 * @summary Submit a quotation request
 */
export const CreateQuotationBody = zod.object({
  customerName: zod.string(),
  customerEmail: zod.string(),
  customerPhone: zod.string().nullish(),
  company: zod.string().nullish(),
  productIds: zod.array(zod.number()),
  message: zod.string(),
  quantity: zod.string().nullish(),
});

/**
 * @summary Update quotation status
 */
export const UpdateQuotationParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateQuotationBody = zod.object({
  status: zod.enum(["pending", "reviewed", "responded", "closed"]).optional(),
});

export const UpdateQuotationResponse = zod.object({
  id: zod.number(),
  customerName: zod.string(),
  customerEmail: zod.string(),
  customerPhone: zod.string().nullish(),
  company: zod.string().nullish(),
  productIds: zod.array(zod.number()),
  message: zod.string(),
  quantity: zod.string().nullish(),
  status: zod.enum(["pending", "reviewed", "responded", "closed"]),
  createdAt: zod.coerce.date(),
});