# lib/api-zod/src/generated/types/categoryWithCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Corporate Gifts Inventory API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryWithCount {
  slug: string;
  name: string;
  productCount: number;
}