# lib/api-zod/src/generated/types/createQuotationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Corporate Gifts Inventory API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQuotationBody {
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  company?: string | null;
  productIds: number[];
  message: string;
  quantity?: string | null;
}