# lib/api-zod/src/generated/types/productStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Corporate Gifts Inventory API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductStats {
  total: number;
  visible: number;
  hidden: number;
  inStock: number;
  lowStock: number;
  outOfStock: number;
  negotiable: number;
  totalCategories: number;
}