# lib/api-zod/src/generated/types/updateProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Corporate Gifts Inventory API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProductBodyStockStatus } from "./updateProductBodyStockStatus";

export interface UpdateProductBody {
  name?: string;
  description?: string;
  category?: string;
  categorySlug?: string;
  imageUrl?: string | null;
  basePrice?: number;
  discountPrice?: number | null;
  minNegotiationPrice?: number | null;
  isNegotiable?: boolean;
  isVisible?: boolean;
  stockStatus?: UpdateProductBodyStockStatus;
  notes?: string | null;
  bulkPricingNotes?: string | null;
}