# lib/api-zod/src/generated/types/updateQuotationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Corporate Gifts Inventory API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateQuotationBodyStatus } from "./updateQuotationBodyStatus";

export interface UpdateQuotationBody {
  status?: UpdateQuotationBodyStatus;
}